pyproject.toml
[project]
name = "pcjr-asm-game-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyserial>=3.5",
]
